+++ pyas2lib/__init__.py
@@ -1,22 +1,22 @@
 from __future__ import absolute_import
+# from pyas2lib.as2 import DIGEST_ALGORITHMS, ENCRYPTION_ALGORITHMS,\
+#     MDN_CONFIRM_TEXT, MDN_FAILED_TEXT, Partner, Organization, Message, MDN
 import sys
-from pyas2lib.as2 import DIGEST_ALGORITHMS, ENCRYPTION_ALGORITHMS,\
-    MDN_CONFIRM_TEXT, MDN_FAILED_TEXT, Partner, Organization, Message, MDN
 
 VERSION = (1, 0, 0)
-__version__ = VERSION
-__versionstr__ = '.'.join(map(str, VERSION))
+__version__ = '.'.join(map(str, VERSION))
+
 
 __all__ = [
     'VERSION',
-    'DIGEST_ALGORITHMS',
-    'ENCRYPTION_ALGORITHMS',
-    'MDN_CONFIRM_TEXT',
-    'MDN_FAILED_TEXT',
-    'Partner',
-    'Organization',
-    'Message',
-    'MDN'
+    # 'DIGEST_ALGORITHMS',
+    # 'ENCRYPTION_ALGORITHMS',
+    # 'MDN_CONFIRM_TEXT',
+    # 'MDN_FAILED_TEXT',
+    # 'Partner',
+    # 'Organization',
+    # 'Message',
+    # 'MDN'
 ]
 
 if (2, 7) <= sys.version_info < (3, 2):