+++ tests/test_basic.py
@@ -6,14 +6,14 @@ class TestBasic(PYAS2TestCase):
 
     def setUp(self):
         self.org = as2.Organization(
-            as2_id='some_organization',
+            as2_name='some_organization',
             sign_key=self.private_key,
             sign_key_pass='test'.encode('utf-8'),
             decrypt_key=self.private_key,
             decrypt_key_pass='test'.encode('utf-8')
         )
         self.partner = as2.Partner(
-            as2_id='some_partner',
+            as2_name='some_partner',
             verify_cert=self.public_key,
             encrypt_cert=self.public_key
         )